feat(admin): add transactions to admin page
Adds a new section to the admin page to view and filter transactions.
The admin page now also takes a `db.Queries` argument to access transaction data.
Transaction amounts are now stored as `REAL` (float64) instead of `INTEGER` (int64) to support decimal values and are divided by 1,000,000 when parsed.

+++ internal/db/migrations/003_transactions.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS transactions (
     user_id TEXT NOT NULL,
     wallet_id TEXT NOT NULL,
     resource TEXT NOT NULL,
-    amount INTEGER NOT NULL,
+    amount REAL NOT NULL,
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES proxy_users(id) ON DELETE CASCADE,
     FOREIGN KEY (wallet_id) REFERENCES wallets(id) ON DELETE CASCADE
